review: fixes and rearrangements
